check author id

+++ Blogplace.Web/Domain/Articles.cs
@@ -1,4 +1,5 @@
-﻿using Blogplace.Web.Infrastructure.Database;
+﻿using Blogplace.Web.Auth;
+using Blogplace.Web.Infrastructure.Database;
 using MediatR;
 
 namespace Blogplace.Web.Domain;
@@ -12,51 +13,62 @@ public class Article(Guid id, string title, string content, Guid authorId)
     public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
 
     public Guid AuthorId { get; } = authorId;
-
 }
+
+public record ArticleDto(Guid Id, string Title, string Content, DateTime CreatedAt, DateTime UpdatedAt, Guid AuthorId);
+public record ArticleSmallDto(Guid Id, string Title, DateTime CreatedAt, DateTime UpdatedAt, Guid AuthorId);
 
 public record CreateArticleResponse(Guid Id);
 public record CreateArticleRequest(string Title, string Content) : IRequest<CreateArticleResponse>;
-public class CreateArticleRequestHandler(IArticlesRepository repository) : IRequestHandler<CreateArticleRequest, CreateArticleResponse>
+public class CreateArticleRequestHandler(SessionStorage sessionStorage, IArticlesRepository repository) : IRequestHandler<CreateArticleRequest, CreateArticleResponse>
 {
     public async Task<CreateArticleResponse> Handle(CreateArticleRequest request, CancellationToken cancellationToken)
     {
-        var userId = Guid.NewGuid(); //todo
+        var userId = sessionStorage.UserId;
         var article = new Article(Guid.NewGuid(), request.Title, request.Content, userId);
         await repository.Add(article);
         return new CreateArticleResponse(article.Id);
     }
 }
 
-public record GetArticleResponse(Article article); //todo dto
+public record GetArticleResponse(ArticleDto article);
 public record GetArticleRequest(Guid Id) : IRequest<GetArticleResponse>;
 public class GetArticleRequestHandler(IArticlesRepository repository) : IRequestHandler<GetArticleRequest, GetArticleResponse>
 {
     public async Task<GetArticleResponse> Handle(GetArticleRequest request, CancellationToken cancellationToken)
     {
         var result = await repository.Get(request.Id);
-        return new GetArticleResponse(result);
+        var dto = new ArticleDto(result.Id, result.Title, result.Content, result.CreatedAt, result.UpdatedAt, result.AuthorId);
+        return new GetArticleResponse(dto);
     }
 }
 
-public record SearchArticlesResponse(IEnumerable<Article> articles); //todo light dtos
+public record SearchArticlesResponse(IEnumerable<ArticleSmallDto> articles); //todo light dtos
 public record SearchArticlesRequest : IRequest<SearchArticlesResponse>; //todo filters
 public class SearchArticlesRequestHandler(IArticlesRepository repository) : IRequestHandler<SearchArticlesRequest, SearchArticlesResponse>
 {
     public async Task<SearchArticlesResponse> Handle(SearchArticlesRequest request, CancellationToken cancellationToken)
     {
         var results = await repository.Search();
-        return new SearchArticlesResponse(results);
+        var dtos = results.Select(x => new ArticleSmallDto(x.Id, x.Title, x.CreatedAt, x.UpdatedAt, x.AuthorId));
+        return new SearchArticlesResponse(dtos);
     }
 }
 
 public record UpdateArticleRequest(Guid Id, string? NewTitle, string? NewContent) : IRequest;
-public class UpdateArticleRequestHandler(IArticlesRepository repository) : IRequestHandler<UpdateArticleRequest>
+public class UpdateArticleRequestHandler(SessionStorage sessionStorage, IArticlesRepository repository) : IRequestHandler<UpdateArticleRequest>
 {
     public async Task Handle(UpdateArticleRequest request, CancellationToken cancellationToken)
     {
         var isChanged = false;
+        //todo get only author of article
         var article = await repository.Get(request.Id);
+
+        if (article.AuthorId != sessionStorage.UserId) 
+        {
+            throw new ArgumentException("Requester is not author of article");
+        }
+
         if(request.NewTitle != null) 
         {
             article.Title = request.NewTitle;
@@ -77,10 +89,17 @@ public class UpdateArticleRequestHandler(IArticlesRepository repository) : IRequ
 }
 
 public record DeleteArticleRequest(Guid Id) : IRequest;
-public class DeleteArticleRequestHandler(IArticlesRepository repository) : IRequestHandler<DeleteArticleRequest>
+public class DeleteArticleRequestHandler(SessionStorage sessionStorage, IArticlesRepository repository) : IRequestHandler<DeleteArticleRequest>
 {
     public async Task Handle(DeleteArticleRequest request, CancellationToken cancellationToken)
     {
+        //todo get only author of article
+        var article = await repository.Get(request.Id);
+        if (article.AuthorId != sessionStorage.UserId)
+        {
+            throw new ArgumentException("Requester is not author of article");
+        }
+
         await repository.Delete(request.Id);
     }
 }
